Properly separated the definition of the SMILES and SMARTS utility functions from the definition of MCP servers.  This fixes the problem of composing unqiue tool servers while maximizing code reuse.
