+++ charge/servers/SMILES_utils.py
@@ -1,0 +1,113 @@
+################################################################################
+## Copyright 2025 Lawrence Livermore National Security, LLC. and Binghamton University.
+## See the top-level LICENSE file for details.
+##
+## SPDX-License-Identifier: Apache-2.0
+################################################################################
+
+from mcp.server.fastmcp import FastMCP
+try:
+    from rdkit import Chem
+    from rdkit.Chem import AllChem, Descriptors
+    from rdkit.Contrib.SA_Score import sascorer
+except ImportError:
+    raise ImportError("Please install the rdkit package to use this module.")
+
+from loguru import logger
+import logging
+
+logging.basicConfig(level=logging.DEBUG)
+
+def canonicalize_smiles(smiles: str) -> str:
+    """
+    Canonicalize a SMILES string. Returns the canonical SMILES.
+    If the SMILES is invalid, returns "Invalid SMILES".
+
+    Args:
+        smiles (str): The input SMILES string.
+    Returns:
+        str: The canonicalized SMILES string.
+    """
+    try:
+        logger.info(f"Canonicalizing SMILES: {smiles}")
+        mol = Chem.MolFromSmiles(smiles)
+        return Chem.MolToSmiles(mol)
+    except Exception as e:
+        return "Invalid SMILES"
+#        return smiles
+
+# Persistent counter to demonstrate statefulness
+SMILES_VERIFICATION_COUNTER = 0
+
+def verify_smiles(smiles: str) -> bool:
+    """
+    Verify if a SMILES string is valid. Returns True if valid, False otherwise.
+
+    Args:
+        smiles (str): The input SMILES string.
+    Returns:
+        bool: True if the SMILES is valid, False otherwise.
+    """
+    try:
+        global SMILES_VERIFICATION_COUNTER
+        SMILES_VERIFICATION_COUNTER += 1
+        logger.info(
+            f"Verifying SMILES: {smiles} used {SMILES_VERIFICATION_COUNTER} times"
+        )
+        Chem.MolFromSmiles(smiles)
+        return True
+    except Exception as e:
+        return False
+
+def get_synthesizability(smiles: str) -> float:
+    """
+    Calculate the synthesizability of a molecule given its SMILES string.
+    Values range from 1.0 (highly synthesizable) to 10.0 (not synthesizable).
+
+    Args:
+        smiles (str): The input SMILES string.
+    Returns:
+        float: The synthesizability score.
+    """
+    try:
+        # logger.info(f"Calculating synthesizability for SMILES: {smiles}")
+        mol = Chem.MolFromSmiles(smiles)
+        if mol is None:
+            logger.warning("Invalid SMILES string or molecule could not be created.")
+            return 10.0  # Default value for invalid SMILES
+        score = sascorer.calculateScore(mol)
+        logger.info(f"Synthesizability score for SMILES {smiles}: {score}")
+        return score
+    except Exception as e:
+        logger.error(f"Error creating molecule from SMILES: {e}")
+        return 10.0
+
+database_of_smiles = []
+
+def known_smiles(smiles: str) -> bool:
+    """
+    Check if a SMILES string is already known.
+
+    Args:
+        smiles (str): The input SMILES string.
+    Returns:
+        bool: True if the SMILES is known to this MCP server, False otherwise.
+    """
+
+    try:
+        logger.info(f"Checking if SMILES is known: {smiles}")
+        mol = Chem.MolFromSmiles(smiles)
+        smiles = Chem.MolToSmiles(mol, isomericSmiles=True)
+
+        if smiles in database_of_smiles:
+            return True
+        else:
+            database_of_smiles.append(smiles)
+            logger.info(f"SMILES not found in the database: {smiles}")
+            return False
+
+    except Exception as e:
+        return False
+
+
+
